Add compiler option to omit type specs when generating
This is since the erlang compiler expects to see referenced
messages/records before the reference to them, which we currently do
not guarantee. Also, there's the possible case of recursively
containing messages/records, for which we cannot find a good
generation order until the erlang compiler accepts recursively defined
types.