Generate code for [optional] message verification
The {verify,How} to gpb_compile:file specifies whether to generate
code that type-verifies always or optionally.

if How == always, the message is always unconditionally type-verified
on encoding in GeneratedCode:encode_msg/1,2.

if How == optionally, an option for encoding may be specified to the
GeneratedCode:encode_msg/2. The option to specify is {verify,boolean()}.