Refactor the code again while keeping binary patterns optimized
The huge `d' decoding function is now split across
several decoding functions, all still getting optimized.

Also, there is a new field in the #field record: is_packed,
(previously this was an option in a list of options) so we
can quickly find out if a field is packed or not.