Verify wiretype vs expected on decoding
When decoding a field, verify that the received wiretype matches what
the field's definition says it should be. Skip received field on
mismatch. The purpose is to avoid decoding values of the wrong type.

Previously decoding was based only on the field's definition, so
upon receiving data with another wiretype, the decoder would likely
crash (sooner or later) or unpack gibberish.