Make names separated by dots rather than underscore
Make names separated by dots instead of underscores.
Originally, I went for underscores to avoid having
to quote the atoms, but I believe it is better to go
for dots as separators since (a) it causes least surprise
and (b) since common practice seems to be to spell message
names capitalized thus one often needs to quote it anyway.