Add an option to generate NIF code for message decoding
The NIF code calls the c++ code generated by Google's protobuf library
to do the actual decoding. Then, the values are transferred to an
Erlang data structure (tuples/records), which is then returned.

The reason is speed, early measurements indicate that decoding with
the NIF is 2 to 3 times faster than the non-NIF version.