Fix compilation-to-binary performance regression
The performance regression was clearly visible with the QuickCheck
tests, which took about 45 times as long to run as previously,
due to an in-retrospect-not-so-cleverly programmed iteration over a
binary. Lots of garbage was created and collected it seems.

The regression appeared in 6eb1488ae09048a9b662c2ed1f768bf5a6e0d23f