Implement name resolution, searching outer scopes
The name resolution was a never very well implemented, and only worked
for most cases, but not when name resolution needed to search outer
scopes to find a name.

This is more or less a complete rewrite and cleanup/refactorisation of
the name resolution process.

Thanks to Heiner Bunjes for drawing attention to it.