Use erl_syntax and erl_syntax_lib functions
Drop the usage of the ad-hoc term-traverser,
in favour of erl_syntax and erl_syntax_lib functions.

The key factor to be able to use erl_syntax and erl_syntax_lib
was to not mix abstract parse-trees with parse-trees. This was
made possible by moving parse-tree manipulations from compile-time
to run-time.