Code generation for decoders: convert the initial msg
The msg0 function, which was only used in the pass-as-record case, is
now converted to the code generator parse-transform. As a side-effect
of this, it has changed slightly. To the better, I think.

Previously: for messages containing inner messages, the generated code
looked like:

    #outer{f = #inner_msg{}}

But now the generated code looks just like:

    #outer{}

The #inner_msg expansion was only done for the pass-as-record
case, not the pass-as-params case, and was not really necessary
anyway.