Refactor codegen transform: keep erl_syntax tree format
Retain the erl_syntax parse-tree format for as long as possible,
revert to the erl_parse parse-tree format only as the last
step. Previously, this was done early and often, now it is either
deferred to the latest possible moment (for gpb_codegen:mk_fn)
or not needed at all (for gpb_codegen:format_fn).

The intention of keeping the erl_syntax parse-tree format, is to be
able to make use of the erl_syntax:text() feature, in order to make an
attempt at generating code for maps instead of records.