For eunit tests, add a {i,"../include"} compiler opt
The reason for this is as follows:
* In test, I have a symlink to gpb_eqc.erl
  (I find this is useful since rebar eunit (or make test) will then
  also run the quickcheck tests. Sure, it takes a bit longer, but it
  has found errors that would otherwise have been very hard to
  discover.)
* Up until recently, the gpb_eqc.erl has had its own copy of the field
  record, but that is obviously not good.
* I also run it regularly from Jenkins, but then the setup is slightly
  different.

So in either case, gpb_eqc.erl ouhgt to include gpb.hrl.
Quesion is how to get paths right in all setups (symlinked, Jenkins)

With an include path to the include directory, I can have just
-include("gpb.hrl") in gpb_eqc.erl, it can be symlinked to from
gpb/test, or it can be executed independently, if I properly set
include paths up there. All good.