Update README.md wrt maps
One of the unsettled issues for maps was encoding of non-present
optional fields, this is now defined, hence no longer an unsettled issue.

The other unsettled issue was whether or not to include some kind of id
or type element in maps, but I've not heard of any demand to have that,
and also don't think it is necessary, so I removed that as an unsettled
issue. Hence no unsettled issues remains.