Add an option for maps to omit non-set optional fields
The option {maps_unset_optional, omitted | present_undefined}
(or cmd line otp -maps_unset_optional omitted | present_undefined)
defines the internal representation for optional fields when the value
is not set: whether it is omitted, or present with the value undefined.
This is only relevant when compiling wth the maps (-maps) option.