Run nif tests in a separate vm
The nif code also brings in the protobuf dynamic library as a library
dynamic dependency, and the protobuf library has special hooks to be
called when the library is load and unload.

With nifs in Erlang, one cannot control unloading of the nif, and thus
cannot control when unloading of the protobuf library happens. I've seen
cases where rapid unloading and reloading of nifs with new proto defs
cause segvs (in strcmp), and from the stack trace, it looks like the
protobuf library is unloading at the time when (or probably while) the
next nif was being loaded. Since the loaded/linked protobuf dynamic
library is a shared resource, if unloading and (re)loading happens in
different os threads, then there might well be trouble.

For code hot upgrade scenarios, this probably means a risk.

So, to make the eunit tests stable, run the nifs in a separate vm.