Replace calls to maps module with matching exprs
For example, now generate code like this:

     case M of
         #{field := Var} -> ...
	 _ -> ...
     end

instead of previously:

     case maps:find(field, M) of
         {ok, Var} -> ...
         error -> ...
     end

The reason for this is performance: Benchmarking shows it appears to
yield a 2-7% speedup on decoding (negligible for encoding).

Matching emits a match beam op, whereas calling the maps module involves
the usual procedure of saving live values on the stack, and restoring
them after the call has returned, just as for any remote call (ie to an
external module). So it saves a few operations for each such operation
plus some garbage building the ok tuple.