Bugfix benchmarking to be more deterministic
The proto-bench now runs the benchmarks in a separate process, to get a
deterministic value.  Previously, everything was executed in the same
process.  This cause an unexpected dependency---introduced in f5da940,
released in 3.21.1---where a large number of command line arguments
would influence the measurement results, because the size of the
process' heap was already initially large enough to make a difference.

Due to this, decoding performance appeared to drop, falsely, in 3.21.1,
because the `make benchmarks` target no longer did run with all the 77
`d' messages on the command line.  The measurement figures for these `d'
messages were never included in the figures in the README, but their
mere presence on the command line nevertheless influenced the figures.

However, the values in the README in 3.21.1 were taken previously, so
the false appearance of performance drop was only visible to whoever
manually ran the benchmarks in 3.21.1 or later.