Fix coding bug for proto3 repeated string,bytes,msg fields
In proto3, repeated numeric scalar fields are by default packed.
Previously, gpb defaulted all repeated fields to packed (except
for map<_,_> fields), but this erroneously also defaulted fields
of the following types to packed:

  - string
  - bytes
  - sub messages

These are no longer defaulted to packed.

As a consequence, there were some necessary code changes: the defaulting
to packed can only occur after all files have been parsed, because enums
are to be defaulted to packed, but references are not resolved until all
files have been parsed, so the defaulting-to-packed stage needed to be
deferred. Thus, also the option normalization stage needed to be deferred.

This coding error was uncovered by fix in 8d4187b (first appearing in
3.24.0) where bugs in the separate vm could go undetected.