Bugfix forgotten option alias normalization
When calling gpb_compile:proto_defs/2,3, the options were not properly
normalized wrt aliasing. Now they are.

This had the effect that the options below were not expanded
properly when gpb_compile:proto_defs/2,3 was called. They were only
expanded properly when gpb_compile:file/1,2 or gpb_compile:string/2,3
were called.

- epb_compatibility (expanded to epb_functions plus two more)
- to_msg_defs (normalized into to_proto_defs)

Also, due to this, the call flow was reworked, so that options are not
expanded twice, if called from eg gpb_compile:file, when the
parse-result is passed on to the common stage to compile proto defs.