Fix proto3 type-default handling
Move from handling type-defaults for required fields only,
to handling it for optional fields.

Originally, when implementing proto3, the parser treated proto3 syntax
fields as required, but this was a mistake, and in b25314b (3.24.4),
this was changed and now the the parser treats them as optional.

But the proto3 type-default handling still kicked in only for required
fields. Now it, too, is moved, so that it kicks in for optional, not
required fields, to mimic the change made ealier in the parser.