Bugfix decoding of some msgs with defaults
If an optional message field had a default, and the optional field was
not present when decoding, then the result of the decoding could
sometimes be that the field had the value 'undefined', or that it had
the default value.  Which of these depended on whether the message was
pass_as_record, or pass_as_params, which in turn depends on the types of
other fields in the message.

This has been fixed, so now the value 'undefined' is used, regardless of
internal message pass method employed.