Don't generate unneeded default translators
Previously, some auxiliary code (default translators) were generated
regardless of whether they were needed or not.  This was for simplicity.
The resulting code could then contain unused functions.  To avoid
compiler warnings for these, the -compile({nowarn_unused_function,F/A}).
attribute was used.

However, it turned out even though the compiler was silenced by this,
the xref tool was not.

So the solution was to stop generating unused functions, making an extra
calculation about which such functions would be used and which were not.