Implode param record return values to maps
Eg turn

  fn(..., A, B, C) ->
       #r{a=A, b=B, c=C}.

into

  fn(..., A, B, C) ->
       #{a => A, b => B, c => C}.

This gets a little more complicated for optional fields with
maps_unset_optional = omitted: for this case, optional fields are
checked for a special `undefined' value, and added if the field is
set.

Underscoring of unused variables/params must be a late stage.  In
particular, for maps, it must consider the map creation, update and
matching expressions, and do this on all generations.  But erl_syntax
text nodes makes this difficult, and erl_syntax_lib:annotate_bindings
does not consider any already existing annotations on text nodes.