Bugfix resolving of references for extender fields
When resolving references for `extend', if one of the fields is a
reference to a message or enum in an enclosing package, for example,
then the name resolution was wrongly calculated.

Also fix the case if a message in another package is extended, and the
type is a reference to message or enum in the package that extends.