Add possibility to unit-test hrl file generation
Previously, hrl file generation were more or less only manually
tested.  The option to compile to memory was implemented by
inserting forms representing the records of the generated hrl,
and the form-representation was expected to be equivalent to the
generated hrl file.

The reason for this odd approach was that the epp only supports
preprocessing from a file, and when the code was already in
memory, it seemed like unnecessary work to land it to the file
system as an intermediary step, managing temporary files,
cleanup etc).

Now, gpb implements nepp, a nano erlang preprocessor, which is
capable of just enough supporting the hrl it generates itself.

And it found a bug (fixed in the previous commit.)