Invoke build helpers as arguments to sh
This is so they can run also when there's no x bit set for the files
in build/*

It seems when downloading the package as a tar file from hex,
the build/* files inside don't have any execute permission bit set.

So invoke helpers as "sh build/cmd" instead of just "build/cmd",
so building works also with gpb as a dependency, eg using rebar3.