Remove a no-longer effective opts override
The override was for generating decoders. It was needed during a
transitionary period, but it was forgotten to remove the override when
it was no longer needed.

The override forced code generation of records and pass_as_record, and
handling of for maps and pass_as_params was reworked into
post-generation passes.