Prepare deprecated erlang:get_stacktrace/0 in 21
In Erlang 21, the function erlang:get_stacktrace/0 has been deprecate
and there is new syntax for retrieving the stacktrace:

  try ..
  catch Class:Reason:Stacktrace -> ...
  end

In gpb and in gpb eunit tests, use a -compile(...) attribute to
silence the deprecation warning, but in generated code, do it more
properly by either generating for the old or the new syntax.
It is slightly complicated by the fact that we might be generating
for a different target version than we're running. In this case,
we use the erl_syntax:text/1 to generate the proper syntax.