Fix rebar eunit compiler options
For "rebar eunit", the ConfigOptions got included twice.  In practice,
it usually meant that for older-than-latest versions of Erlang/OTP when
a compatibility option was needed, when running rebar eunit,
there was an error about redefinition of a macro, like:

  src/gpb_<file>.erl:none: redefining macro 'NO_HAVE_XYZ'

This was because for eunit, rebar does erl_opts ++ eunit_compile_opts ++ ...
hence the macro redefinition.