Generate optional stacktrace code
The purpose of this is the use case when generated code is checked in,
for instance when there cannot be dependencies such as in hex_erl,
and yet the code must (as far as possible) compile across a
variety of different versions of Erlang.

There is currently no preprocessor-only approach, so external
support is needed, such as from a build system. However,
we can generate code targeted for some version, where it is
possible to override with a define, if compiling some some
other version.

For Erlang 21 or later, generate code that uses the new stacktrace
syntax, but if the user defines 'GPB_FUNCTION_STACK', it overrides.

For Erlang 20 or earlier, generate code that calls erlang:get_stacktrace/0
syntax, but if the user defines 'GPB_PATTERN_STACK', it overrides.