Bugfix nested 'extend' blocks
The computation of which message to extend was not correct. It could
extend the containing message even when referring to another message.

The error: Given

    message A {
    }

    message B {
        extend A {
            optional B b = 42;
        }
    }

the message B was wrongly extended with a new field b, when A should
have been extended. In this commit, A gets extended with the field b.