Use the OTP_RELEASE macro to determine stacktrace
Use the OTP_RELEASE macro to select at compile-time for the generated
code, whether to acquire the stacktrace as a pattern or as
previously (using erlang:get_stacktrace/0).

The OTP_RELEASE macro appears in Eralng 21, but at the time of this
commit, it is not yet available in any release or even pre-release yet.
So for a while, keep the previous code that checks for overriding macros;
it may still be a Erlang 21-rc1 where there is no OTP_RELEASE and
erlang:get_stacktrace/0 is deprecated.