pbt: convert gpb-eqc to PropEr & add to test suite
This is a starting point for more thorough testing.
I intend on extending https://github.com/tomas-abrahamsson/gpb-eqc
with PropEr's Targeted testing properties.

This commit is rougly corresonding to bd08d56 (= for-4.3.0) of that
repo, including testing of flat oneof and handling that translation
can also occur on top-level.

This adds a test-time dependency from gpb to rebar3_proper and to
proper.

In travis, test on OTP >= 21.0 for now, since this takes time.

To test manually, run either

* rebar3 as test proper; or
* make proper-test

There is no rebar2 support for running the property based tests.