On encoding of proto3 enums, handle 0 as default too
When an enum is encoded in proto3, if it is the default, it must not
be included in the result. The first enum is the enum's default, and
that enum must also be 0. Previously the default value handling was
done only when the enum was set as the atom, but now it is also done
when set as the integer value.