For proto3 on decoding for maps, remove redundant check
On decoding, for maps, when unset_optional=omitted (which is default),
the decoding code look like this:

   S2 = if F1 == '$undef' -> S1;
           true -> S1#{field_name => F1
        end,

It looked like this for both proto2 and proto3, buit for proto3 the
check for '$undef' is unnecessary (for scalar fields), since they always
have a type-default. Hoever sub messages and oneof do not have a
type-default in proto3, so for such fields, there is still a check.