Bugfix decoding of omittted proto3 submsgs for maps
With the option maps, with maps_unset_optional=omitted (which is default
since 4.0.0), for syntax="proto3" there was a bug on decoding where
unset sub messages in particular, got the value 'undefined' but it
should instead have been omitted.

It was a bug only for proto3; for syntax="proto2" the value was omitted
as expected.