Generate service and rpc name introspection functions
Generate:

  fqbin_to_service_name/1
  service_name_to_fqbin/1

  fqbins_to_service_and_rpc_name/2
  service_and_rpc_name_to_fqbins/2

Consider current values for package as well as the use_package option
and any renam options, when translating between names. The term fqbin
here refers to a fully qualified service name, ie with package name
prepended, and as a binary.

The intended use case for these generated functions is grpc.