Fix rebar2/3 descriptor compiling
Previously, the logic was that if the ebin directory didn't exist,
it was rebar3, else it was rebar2.

This worked well when building gpb as a dependency, but was a bit
awkward when developing gpb itself, especially when switching between
rebar3, rebar2 and make: sometimes there was a (possibly empty) ebin
directory, preventing rebar3 from compiling descriptors correctly.

Now, invert it the logic by checking for $REBAR_DEPS_DIR/gpb.
gpb will never be a dependency to itself, so it cannot exist in the old
rebar2 build structure, but with rebar3, $REBAR_DEPS_DIR
will typically be .../_build/lib/gpb where gpb is about to be built,
so set up shell variables accordingly.

Also, when compiling tests, check whether there is a test
sub directory:
- it will exist for instance with rebar3 eunit
- it will not exist for rebar3 compile
- it will not exist when compiling gpb as a hex depdency
  because the hex package does not contain the test/