Make wellknown types available also for syntax="proto2"
It appears the with protoc, well known types are available without any
special -I options also for syntax="proto2". I originally got the
impression this was only the case for syntax="proto3" but it appears to
hold also for proto2.

However, the protobuf 2.x don't know how to do that. It appears this is
only possible since protobuf 3.x or later.