Generate code to encode to a JSON representation
If the JSON option is specified, a to_json/1,2,3 function
is generated. This can turn the internal representation
of a protobuf message into a JSON representation, suitable
for encoding to text with some other JSON library.

There is support for several of the common libraries, and
it is possible to tweak JSON representation details for other
libraries, too.