Add renaming options for options {msg,enum}_typename
One motivator is for google.protobuf.Any, when combined with message
renaming options snake_case and basename, as grpcbox mandates. The
resulting message type will be any)' which collides with the built-in.
type.

So add renaming options

   {rename, {msg_typename, How}}
   {rename, {enum_typename, How}}

That way, the type names (but not the message or enum names) can be
prefixed or suffixed at will to avoid collisions.