Remove an unused decoder step
During the great refactorization of the decoding, before 4.0.0,
generation of decoders was changed to be into two phases: first generate
code only for records, then turn those records to maps depending on
options.

Prior to that, the generation of decoders had to consider targetting
records or maps while generating the code (focusing on more than one
thing at a time).

During the transition, for a brief period, between 832cb63057 and
24d7d699ea, code generation forced records, while still being able to
generate for maps.  In 24d7d699ea it lost the ability to generate for
maps (relying on this happening instead in the second step). But the
forcing of records was still left there by mistake.  Remove it; it is
not needed any more.