Update travis config: -Erlang17
Drop support also for Erlang 17, since it turned out Rebar3 dropped
support for Erlang 17 when Erlang 22 arrived.

So follow suit, and take the opportunity to open for code
simplifications in gpb.

Map expressions arrived in Erlang 17, and the possibility to use
variables as map keys arrived in Erlang 18.

In gpb, the option {maps_oneof, flat} requires Erlang 18, since it
generates expressions with variables as map keys, so dropping support
for Erlang 17 and earlier means that in the code generator, it will
be possible to use all needed map expressions directly. The current
workaround for limitations in Erlang =< 17 relies on erl_syntax text
nodes and expressing it is a level of indirection at every place,
so being able to use map epressions directly will lead to clearer
and more readable code.