Generate NIF code for JSON
This makes it posssible to cross-check the JSON format.
Run a message via the Erlang to_json through Googe's protobuf
JsonStringToMessage and check that we got the message back.
And vice versa in the other direction.

In decoders, feed TrUserData down to the implementing
to_json_msg_<MsgName> and from_json_msg_<MsgName> functions
only if they are implemented in Erlang; don't if implemented as NIFs.