Bugfix flat maps for proto3 with oneof and nifs
For proto3, with flat maps, for oneof fields, for NIF, decoding could
return all oneof field alternatives set, which could clearly not be ok.

The problem was the checking of what field was set for flat maps. Due
to the (wrong) approach of unrolling the oneof field alternatives,
it no longer checked the <field>_case(), but instead checked
m->has_<alterative>() for each oneof field, and due to proto3
defaulting all fields to their type defaults, this was always true.