Robustify finding vendored proto3 well-knowns
Detect the case that there is another app such as gpb-addon
before gpb in the in the Erlang vm's code path.  It can wrongly
think the -addon part indicates a version number, and that gpb
is its name.  Detect that and use the fallback.

In the fallback, verify a bit closer that the heuristically
calculated directory actually exists.