Rework call to deprecated stacktrace
Change the call to the deprecated erlang:get_stacktrace()
into the stacktrace syntax, on Erlang >=21, using a macro,
depending on the presence of the OTP_RELEASE macro (which
also was introduced in Erlang 21)

Previously, instead of checking for OTP_RELEASE, it was
a by including a nowarn_deprecated_function directive
conditionally if teh build Makefile or rebar.config.script
detected OTP 21 or later.  This got rid of the compilation
warning about the function being deprecated, but with xref,
the warning was visible anyway. So use a macro to do
the proper syntax anyway.

For a few alpha releases just before Erlang 21, the
OTP_RELEASE macro was not present, yet still stacktrace syntax
was needed, but no need to support that situation any more.