Add a recursive descent scanner and parser
The problem this aims to solve in the end, is to allow keywords
everywhere. For instance the following is not possible with the
previous solution, because 'rpc' and 'enum' are keywords.
(the nonsensical names in this example are only for illustrating
the problem) [*]:

    message rpc {
      optional enum field1 = 1;
    }
    message enum {
      optional uint32 field1 = 1;
    }

The solution here is to implement a recursive descent parser.

Also implement a scanner, and let it parse binaries, which are
expected to be UTF-8. This ought to use less memory, and feels
more modern.

For now, aim for the same feature set as the previous scanner/parser,
and let it exist in parallel, so it can be possible to just switch
the parser/scanner with an option, for testing purposes

_____
[*] It is not possible to fix the grammar to allow keywords
    at all places where an identifier is expected, it would lead
    to reduce/reduce conflicts.