Fix dialyzer issue for map<_,_> in generated code
With records, if either fields were passed in records (as opposed to
as variables, which analysis often finds preferable for performance
reasons) or if json was to be emitted, then there were dialyzer issues
in the generated code. These are now fixed.

The dialyzer issues were due the fact that map<_,_> fields were stored
during decoding or from-jsoning as dict:dict() instead of as a list of
2-tuples. As finalization, the dicts were converted to tuple lists,
but the intermediary storage as dicts violtated the type spec.

Fix by storing as list of 2-tuples, and move the passing via
dicts to the finalization pass instead.

(The dict are used to handle the case when the same key would occur
more than once in the input-to-be-decoded.)