Fix type specs for translated flat oneof alternatives
For the maps_oneof=flat option, translated oneof alternatives
did not consider (translation-specified) type specs at all.
This is fixed. Example:

   [maps, {maps_oneof,flat},
    {tranalsate_field, [{['Msg',c,a1], [...,{type_spec, "x:t1()"}]}]

The following type was erroneously generated:

   -type 'Msg' :: #{a1 => integer()}.

Now this gets generated:

   -type 'Msg' :: #{a1 => x:t1()}.

Type string and type comment generation is now done in a preparation
phase where info is first gathered. Following this is, a stage where
type strs and comments are either combined (if non-flat oneofs)
or flattened (for flat oneofs).