Return proto defs on version 1 by default
This is when returning proto defs, ie the to_proto_defs option is
specified.

This allows introduction of new non-backwards features in the
proto defs format without breaking legacy uses.

THe plan is for change proto3 occurrence to accomodate for new
proto3 field presence features, since this makes the code more logical
internally. If proto_defs_version 1 is to be returned, the intention is
that it will be possible to convert to and from this format losslessly
as long as no new proto3 field presence features are used.

That is, the new proto3 field presence feature features will require
proto_defs_version 2.

But if no such field presence features are used, the gpb works in a
backwards compatible way by default.