Change proto3 occurrence: 'optional' to 'defaulty'
What changed and why
--------------------

In proto3, change the representation of proto3 fields that are
non-repeated: from occurrence = optional to occurrence = defaulty.

For example:

        syntax = "proto3";
        message Message {
          uint32 field1 = 1;
        }

Previously, this used to get represented as:

        [{{msg,'Message'}, [#?gpb_field{name=field1,
                                        occurrence=optional}]},
         ...]

Now this is represented as below, note how occurrence=defaulty:

        [{{msg,'Message'}, [#?gpb_field{name=field1,
                                        occurrence=defaulty,
         ...]

The purpose is to make room for the upcoming
new field presence handling in proto3, that is experimental
in protobuf 3.12.0: The 'optional' appears to be making a come-back,
so the following is about to become valid:

        syntax = "proto3";
        message Message {
          optional uint32 field1 = 1;
        }

proto_defs_version and backwards compatibility
----------------------------------------------

This changes the internal representation so the latest
proto_defs_version is now 2, and this is the version used internally
when generating an erl source file or a binary.

However, in this commit, when returning proto definitions (if the
to_proto_defs option is specified), it is still returned on
proto_defs_version 1 format.  There are version conversions,
if necessary, on input (eg gpb_compile:proto_defs/2,3) and output
(the to_proto_defs option as well as introspection functions).

Misc other
----------

* In general, now, code can check for occurrence == defaulty,
  and know it is a proto3 field with proto3 (traditional) presence
  semantics.  Previously, the code had to trigger on
  occurrence = optional + message is proto3 (but if the type is
  a sub message, treat it like proto2 anyway)

* Recognize the fact that the map<_,_> pseudo message behave slightly
  different in encoding and decoding.  During analysis, save another
  set of psuedo messages for decoding.