Add support for proto3 optional
In protobuf 3.12.0, experimental support for optional
in proto3 was added.  This commit adds corresponding
support to gpb.

It is now possible to write the following in a .proto

    syntax="proto3";
    message Msg {
      uint32 f1 = 1;
      optional uint32 f2 = 2;
    }

The field f1 still behaves like proto3 fields have previously
been handled.

The field f2 is marked optional, and essentially, it behaves like
if it would have been in a proto2 message:

  - if it is not set at encoding, it is omitted from
    the encoded binary

  - on decoding, if it is not included in the binary to decode,
    then it gets set to undefined (if records) or omitted from
    the map (if maps (unless maps_unset_optional = present_undefined))

When a proto3 message field is marked optional, then it requires proto
defs version 1 and it is not possible to convert the definitions to
proto defs version 1.  The definitons cannot be accurately represented
in that version.

In the descriptors, an extra field is set, proto3_optional,
for marking that this is a proto3 optional field, and synthetic
oneofs are introduced.