Pivot scanners,parsers: the default is now the new
The xrl yrl based one is now the old (with _old suffix).
The new recursive descent is the default (no longer 2 in the file name).

Change unit tests to test the new scanner and parser.

To ease transition to the new parser, in case there should be any
errors in the new parser (have tested fairly thoroughly though),
there are still some ways to run with the old parser instead

* the undocumented option {parser, old | new}
* the environment variable GPB_PARSER=old|new

Note that these options will be removed after the
new parser has been in use for some time!