Parse uninterpreted option blocks
For now, just save them as {uninterpreted, "...approximate text..."}

The text as such ought not to be too strictly interpreted,
some of the original information is currently lost, and it is
not enough to recreate it.

Example:

  message Msg {
    option (my_option) = {
      f1: 'str'
      f2: 3
    }
    optional uint32 f1 = 1;
  }

The main reason for adding this support is to be able to parse a
corpus of proto files with no parsing errors.