Emit presence as a comment to fields' type specs
The commits ccf6d40a58 and 282d2aecda changes presence
indication in type specs to allow it to be "unset"
even for 'required' fields, since at decoding, actual
presence depends on the input data to decode.

This commit adds required, optional etc in comments
in the type specs, to still give some kind of indication
of intention.