Fix the import_fetcher to apply to imports only
The normal use case when not using the file system
is gpb_compile:string with an import_fetcher option.

However, when calling gpb_compile:file/1,2 and also supplying an
{import_fetcher,Fun} option, the Fun was invoked also for the
top-level file, ie for the file argument to gpb_compile:file/1,2.

This commit changes the gpb_compile:file/1,2 to use any
import_fetcher option only for imports, not for its main file
parameter, which makes more sense given the name of the option.

Potential backwards incompatibility, though, for anyone that
relied on the previous somewhat odd behaviour.