Refactor record and type generation
Separate calculation of records and map fields from the
formatting and layout, to get somehat more readable code.

One bug got fixed: The previous implementation sometimes got it wrong
when figuring out whether to add a comma after the field or not.  That
is, it did not always figure out correctly when the rendered field was
the last non-out-commented (map) field or not.  The result was that it
could generate (map) type specs that would not compile.  This has been
fixed.