Fix map type spec assoc for binary keys
This is the same as in 282d2ae but map key type = binary:

Use := only when maps_unset_optional = present_undefined.
Otherwise, for maps_unset_optional = omitted: use => even
when there are required fields, since we cannot guarantee
presence at decoding, since it depends on the input binary
to decode.