During decoding pass an extra param
This is a preparation for being able to preserve unknown fields:
Pass an extra argument, which will be used for field number.

The reason for always passsing it, is that when passing fields
as paramters (often the case) the field params can be kept at the
same positions, so the compiler need not generate any instructions
to shift up or down the field params, these are just left where
they are, and stay put.