Internally, treat include paths as absolute
For relative path {i,Dir} options, the relative path is converted
to an absolute path, and the absolute paths are used when checking
whether an import file has already seen or not. Previously, the
relative path was used in this check.

The paths are normalized in the sense that . and .. are reduced:

   /a/b/c/../foo -> /a/b/foo
   /a/b/./foo    -> /a/b/foo

No attempt is made to dereference symlinks in this process.

For list_io/2, the same relative paths as previously
are returned, so that any generated dependency rules
such as Makefile rules, are relocatable as far as possible.
However when checking whether an import was already seen,
then the only full paths are considered.